led-controller-uservice/requirements.txt
flask==0.12.2
pyserial==3.3
